clean up of SDK
